Made parameters an interface
